backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
google-generativeai==0.8.5
matplotlib==3.9.4
seaborn==0.13.2
